refactor: move combo files to module folders
