+++ src/modules/auth/forms/edit-account.tsx
@@ -7,12 +7,12 @@ import { TFunction, useTranslation } from 'next-i18next'
 import * as yup from 'yup'
 
 import { Container } from '@/components/forms/container'
-import { ClubsCombo } from '@/components/selects/clubs-combo'
-import { RegionsCombo } from '@/components/selects/regions-combo'
-import { UserFootballRolesCombo } from '@/components/selects/user-football-roles-combo'
 import { UpdateUserDto, User } from '@/modules/auth/auth'
+import { ClubsCombo } from '@/modules/clubs/combo'
 import { useClubsList } from '@/modules/clubs/hooks'
+import { RegionsCombo } from '@/modules/regions/combo'
 import { useRegionsList } from '@/modules/regions/hooks'
+import { UserFootballRolesCombo } from '@/modules/user-football-roles/combo'
 import { useUserFootballRolesList } from '@/modules/user-football-roles/hooks'
 import { validateId } from '@/utils/validation-helpers'
