+++ src/modules/countries/combo.tsx
@@ -3,13 +3,13 @@ import { Field } from 'formik'
 import { Autocomplete } from 'formik-mui'
 import { useTranslation } from 'next-i18next'
 
-import { PlayerPositionDto } from '@/modules/player-positions/types'
+import { IComboProps } from '@/types/combo'
 
-import { IComboProps } from './types'
+import { CountryDto } from './types'
 
-interface IPlayersPositionComboProps extends IComboProps<PlayerPositionDto> {}
+interface ICountriesComboProps extends IComboProps<CountryDto> {}
 
-export const PlayersPositionCombo = ({
+export const CountriesCombo = ({
   data,
   name,
   label,
@@ -17,7 +17,7 @@ export const PlayersPositionCombo = ({
   size,
   error,
   helperText,
-}: IPlayersPositionComboProps) => {
+}: ICountriesComboProps) => {
   const { t } = useTranslation()
 
   return (
@@ -27,14 +27,14 @@ export const PlayersPositionCombo = ({
       multiple={multiple}
       id={name}
       size={size}
-      options={[0, ...data.map(position => position.id)]}
+      options={[0, ...data.map(country => country.id)]}
       getOptionLabel={(option: number) => {
         if (option === 0) {
           return ''
         }
-        const position = data.find(p => p.id === option)
-        if (position) {
-          return position.name
+        const country = data.find(c => c.id === option)
+        if (country) {
+          return country.name
         }
         return t('NONE')
       }}
@@ -44,8 +44,8 @@ export const PlayersPositionCombo = ({
           {...params}
           error={error}
           helperText={helperText}
-          label={label || t('PRIMARY_POSITION')}
-          placeholder={label || t('PRIMARY_POSITION')}
+          label={label || t('COUNTRIES')}
+          placeholder={label || t('COUNTRIES')}
         />
       )}
     />
