+++ src/modules/players/forms/filter.tsx
@@ -6,17 +6,17 @@ import { useTranslation } from 'next-i18next'
 
 import { Container } from '@/components/forms/container'
 import { FilterFormActions } from '@/components/forms/filter-form-actions'
-import { CompetitionGroupsCombo } from '@/components/selects/competition-groups-combo'
-import { CompetitionsCombo } from '@/components/selects/competitions-combo'
-import { CountriesCombo } from '@/components/selects/countries-combo'
-import { FootedSelect } from '@/components/selects/footed-select'
-import { PlayersPositionCombo } from '@/components/selects/player-positions-combo'
-import { TeamsCombo } from '@/components/selects/teams-combo'
+import { CompetitionGroupsCombo } from '@/modules/competition-groups/combo'
 import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
+import { CompetitionsCombo } from '@/modules/competitions/combo'
 import { CompetitionBasicDataDto } from '@/modules/competitions/types'
+import { CountriesCombo } from '@/modules/countries/combo'
 import { CountryDto } from '@/modules/countries/types'
+import { PlayersPositionCombo } from '@/modules/player-positions/combo'
 import { PlayerPositionDto } from '@/modules/player-positions/types'
+import { FootedSelect } from '@/modules/players/footed-select'
 import { PlayersFiltersDto } from '@/modules/players/types'
+import { TeamsCombo } from '@/modules/teams/combo'
 import { TeamBasicDataDto } from '@/modules/teams/types'
 
 const StyledCheckboxContainer = styled('div')(() => ({
