+++ src/modules/regions/combo.tsx
@@ -3,13 +3,13 @@ import { Field } from 'formik'
 import { Autocomplete } from 'formik-mui'
 import { useTranslation } from 'next-i18next'
 
-import { ClubBasicDataDto } from '@/modules/clubs/types'
+import { IComboProps } from '@/types/combo'
 
-import { IComboProps } from './types'
+import { RegionDto } from './types'
 
-interface IClubsComboProps extends IComboProps<ClubBasicDataDto> {}
+interface IRegionsComboProps extends IComboProps<RegionDto> {}
 
-export const ClubsCombo = ({
+export const RegionsCombo = ({
   data,
   name,
   label,
@@ -17,7 +17,7 @@ export const ClubsCombo = ({
   size,
   error,
   helperText,
-}: IClubsComboProps) => {
+}: IRegionsComboProps) => {
   const { t } = useTranslation()
 
   return (
@@ -27,14 +27,14 @@ export const ClubsCombo = ({
       multiple={multiple}
       id={name}
       size={size}
-      options={[0, ...data.map(club => club.id)]}
+      options={[0, ...data.map(country => country.id)]}
       getOptionLabel={(option: number) => {
         if (option === 0) {
           return ''
         }
-        const club = data.find(c => c.id === option)
-        if (club) {
-          return club.name
+        const region = data.find(r => r.id === option)
+        if (region) {
+          return region.name
         }
         return t('NONE')
       }}
@@ -44,8 +44,8 @@ export const ClubsCombo = ({
           {...params}
           error={error}
           helperText={helperText}
-          label={label || t('CLUB')}
-          placeholder={label || t('CLUB')}
+          label={label || t('REGIONS')}
+          placeholder={label || t('REGIONS')}
         />
       )}
     />
