+++ src/modules/teams/forms/create.tsx
@@ -5,12 +5,12 @@ import { useTranslation } from 'next-i18next'
 
 import { Container } from '@/components/forms/container'
 import { MainFormActions } from '@/components/forms/main-form-actions'
-import { ClubsCombo } from '@/components/selects/clubs-combo'
-import { CompetitionGroupsCombo } from '@/components/selects/competition-groups-combo'
-import { CompetitionsCombo } from '@/components/selects/competitions-combo'
 import { useAlertsState } from '@/context/alerts/useAlertsState'
+import { ClubsCombo } from '@/modules/clubs/combo'
 import { ClubBasicDataDto } from '@/modules/clubs/types'
+import { CompetitionGroupsCombo } from '@/modules/competition-groups/combo'
 import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
+import { CompetitionsCombo } from '@/modules/competitions/combo'
 import { CompetitionBasicDataDto } from '@/modules/competitions/types'
 import { CreateTeamDto } from '@/modules/teams/types'
