+++ src/modules/teams/forms/filter.tsx
@@ -6,15 +6,15 @@ import { useTranslation } from 'next-i18next'
 
 import { Container } from '@/components/forms/container'
 import { FilterFormActions } from '@/components/forms/filter-form-actions'
-import { ClubsCombo } from '@/components/selects/clubs-combo'
-import { CompetitionGroupsCombo } from '@/components/selects/competition-groups-combo'
-import { CompetitionsCombo } from '@/components/selects/competitions-combo'
-import { CountriesCombo } from '@/components/selects/countries-combo'
-import { RegionsCombo } from '@/components/selects/regions-combo'
+import { ClubsCombo } from '@/modules/clubs/combo'
 import { ClubBasicDataDto, ClubsFiltersDto } from '@/modules/clubs/types'
+import { CompetitionGroupsCombo } from '@/modules/competition-groups/combo'
 import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
+import { CompetitionsCombo } from '@/modules/competitions/combo'
 import { CompetitionBasicDataDto } from '@/modules/competitions/types'
+import { CountriesCombo } from '@/modules/countries/combo'
 import { CountryDto } from '@/modules/countries/types'
+import { RegionsCombo } from '@/modules/regions/combo'
 import { RegionDto } from '@/modules/regions/types'
 
 const StyledCheckboxContainer = styled('div')(() => ({
