+++ src/modules/user-football-roles/combo.tsx
@@ -3,21 +3,22 @@ import { Field } from 'formik'
 import { Autocomplete } from 'formik-mui'
 import { useTranslation } from 'next-i18next'
 
-import { TeamBasicDataDto } from '@/modules/teams/types'
+import { IComboProps } from '@/types/combo'
 
-import { IComboProps } from './types'
+import { UserFootballRoleDto } from './types'
 
-interface ITeamsComboProps extends IComboProps<TeamBasicDataDto> {}
+interface IUserFootballRolesComboProps
+  extends IComboProps<UserFootballRoleDto> {}
 
-export const TeamsCombo = ({
+export const UserFootballRolesCombo = ({
   data,
   name,
   label,
-  multiple,
   size,
+  multiple,
   error,
   helperText,
-}: ITeamsComboProps) => {
+}: IUserFootballRolesComboProps) => {
   const { t } = useTranslation()
 
   return (
@@ -27,14 +28,14 @@ export const TeamsCombo = ({
       multiple={multiple}
       id={name}
       size={size}
-      options={[0, ...data.map(team => team.id)]}
+      options={[0, ...data.map(role => role.id)]}
       getOptionLabel={(option: number) => {
         if (option === 0) {
           return ''
         }
-        const team = data.find(c => c.id === option)
-        if (team) {
-          return team.name
+        const role = data.find(r => r.id === option)
+        if (role) {
+          return role.name
         }
         return t('NONE')
       }}
@@ -44,8 +45,8 @@ export const TeamsCombo = ({
           {...params}
           error={error}
           helperText={helperText}
-          label={label || t('TEAM')}
-          placeholder={label || t('TEAM')}
+          label={label || t('FOOTBALL_ROLE')}
+          placeholder={label || t('FOOTBALL_ROLE')}
         />
       )}
     />
